+++ mclib/pqueue.h
@@ -18,10 +18,10 @@
 // Structure and Class Definitions
 
 typedef struct _PQNode {
-	long			key;			// sort value
-	unsigned long	id;				// hash value for this map position
-	long			row;			// HB-specific
-	long			col;			// HB-specific
+	int32_t         key;			// sort value
+	uint32_t        id;				// hash value for this map position
+	int32_t         row;			// HB-specific
+	int32_t         col;			// HB-specific
 } PQNode;
 
 class PriorityQueue {
@@ -29,13 +29,13 @@ class PriorityQueue {
 	protected:
 
 		PQNode*		pqList;
-		long		maxItems;
-		long		numItems;
-		long		keyMin;
+		int32_t     maxItems;
+		int32_t     numItems;
+		int32_t     keyMin;
 
-		void downHeap (long curIndex);
+		void downHeap (int curIndex);
 
-		void upHeap (long curIndex);
+		void upHeap (int curIndex);
 
 	public:
 
@@ -48,27 +48,27 @@ class PriorityQueue {
 			init();
 		}
 
-		long init (long maxItems, long keyMinValue = -2000000);
+		int init (int maxItems, int keyMinValue = -2000000);
 
-		long insert (PQNode& item);
+		int insert (PQNode& item);
 
 		void remove (PQNode& item);
 
-		void change (long itemIndex, long newValue);
+		void change (int itemIndex, int newValue);
 
-		long find (long id);
+		int find (unsigned int id);
 		
 		void clear (void) {
 			numItems = 0;
 		}
 
-		long getNumItems (void) { return(numItems); }
+		int getNumItems (void) { return(numItems); }
 		
 		bool isEmpty (void) {
 			return(numItems == 0);
 		}
 		
-		PQNode* getItem (long itemIndex) {
+		PQNode* getItem (int itemIndex) {
 			return(&pqList[itemIndex]);
 		}
